Refactor LoadView component by separating MaterialSection and SuggestedVehicleSection into their own files; enhance styling in loadView.scss
