+++ src/components/organisms/logistics/orders/CreateOrderVieww/components/LoadView/loadView.scss
@@ -4,6 +4,27 @@
     .subTitle {
         font-size: 1.25rem;
         font-weight: 600;
-        margin-bottom: 0.5rem;
+        margin-bottom: -0.5rem;
+    }
+
+    .addButton {
+        margin: 12px 0;
+        margin-right: 1rem;
+        border: none;
+        font-weight: 600;
+        font-size: 1rem;
+        transition: all 0.3s ease;
+        box-shadow: none;
+        width: fit-content;
+        align-self: flex-end;
+
+        &:hover {
+            opacity: 0.8;
+        }
+    }
+
+    .divider {
+        border: none;
+        border-top: 1px solid $dark-gray;
     }
 }
